Refactor analyzer
